Added Dejal target & scheme
Please leave the “Turn Touch Dejal” target and build scheme in the project, so I’m able to run the app while working on it. It won’t interfere with anything; you can just build the normal “Turn Touch Mac” target. Select that as the active scheme in the toolbar.